+++ src/providers/auth0.ts
@@ -1,10 +1,9 @@
 import {
-	AuthorizationCodeAuthorizationURL,
-	AuthorizationCodeTokenRequestContext,
-	RefreshRequestContext,
-	TokenRevocationRequestContext
-} from "@oslojs/oauth2";
-import { sendTokenRequest, sendTokenRevocationRequest } from "../request.js";
+	createOAuth2Request,
+	encodeBasicCredentials,
+	sendTokenRequest,
+	sendTokenRevocationRequest
+} from "../request.js";
 
 import type { OAuth2Tokens } from "../oauth2.js";
 
@@ -26,31 +25,44 @@ export class Auth0 {
 		this.redirectURI = redirectURI;
 	}
 
-	public createAuthorizationURL(state: string): AuthorizationCodeAuthorizationURL {
-		const url = new AuthorizationCodeAuthorizationURL(this.authorizationEndpoint, this.clientId);
-		url.setRedirectURI(this.redirectURI);
-		url.setState(state);
+	public createAuthorizationURL(state: string, scopes: string[]): URL {
+		const url = new URL(this.authorizationEndpoint);
+		url.searchParams.set("client_id", this.clientId);
+		url.searchParams.set("state", state);
+		url.searchParams.set("scope", scopes.join(" "));
+		url.searchParams.set("redirect_uri", this.redirectURI);
 		return url;
 	}
 
 	public async validateAuthorizationCode(code: string): Promise<OAuth2Tokens> {
-		const context = new AuthorizationCodeTokenRequestContext(code);
-		context.authenticateWithHTTPBasicAuth(this.clientId, this.clientSecret);
-		context.setRedirectURI(this.redirectURI);
-		const tokens = await sendTokenRequest(this.tokenEndpoint, context);
+		const body = new URLSearchParams();
+		body.set("grant_type", "authorization_code");
+		body.set("code", code);
+		body.set("redirect_uri", this.redirectURI);
+		const request = createOAuth2Request(this.tokenEndpoint, body);
+		const encodedCredentials = encodeBasicCredentials(this.clientId, this.clientSecret);
+		request.headers.set("Authorization", `Basic ${encodedCredentials}`);
+		const tokens = await sendTokenRequest(request);
 		return tokens;
 	}
 
 	public async refreshAccessToken(refreshToken: string): Promise<OAuth2Tokens> {
-		const context = new RefreshRequestContext(refreshToken);
-		context.authenticateWithHTTPBasicAuth(this.clientId, this.clientSecret);
-		const tokens = await sendTokenRequest(this.tokenEndpoint, context);
+		const body = new URLSearchParams();
+		body.set("grant_type", "refresh_token");
+		body.set("refresh_token", refreshToken);
+		const request = createOAuth2Request(this.tokenEndpoint, body);
+		const encodedCredentials = encodeBasicCredentials(this.clientId, this.clientSecret);
+		request.headers.set("Authorization", `Basic ${encodedCredentials}`);
+		const tokens = await sendTokenRequest(request);
 		return tokens;
 	}
 
-	public async revokeToken(refreshToken: string): Promise<void> {
-		const context = new TokenRevocationRequestContext(refreshToken);
-		context.authenticateWithHTTPBasicAuth(this.clientId, this.clientSecret);
-		await sendTokenRevocationRequest(this.tokenRevocationEndpoint, context);
+	public async revokeToken(token: string): Promise<void> {
+		const body = new URLSearchParams();
+		body.set("token", token);
+		const request = createOAuth2Request(this.tokenRevocationEndpoint, body);
+		const encodedCredentials = encodeBasicCredentials(this.clientId, this.clientSecret);
+		request.headers.set("Authorization", `Basic ${encodedCredentials}`);
+		await sendTokenRevocationRequest(request);
 	}
 }