+++ src/providers/lichess.ts
@@ -1,8 +1,5 @@
-import {
-	AuthorizationCodeAuthorizationURL,
-	AuthorizationCodeTokenRequestContext
-} from "@oslojs/oauth2";
-import { sendTokenRequest } from "../request.js";
+import { createS256CodeChallenge } from "../oauth2.js";
+import { createOAuth2Request, sendTokenRequest } from "../request.js";
 
 import type { OAuth2Tokens } from "../oauth2.js";
 
@@ -17,26 +14,29 @@ export class Lichess {
 		this.clientId = clientId;
 		this.redirectURI = redirectURI;
 	}
-
-	public createAuthorizationURL(
-		state: string,
-		codeVerifier: string
-	): AuthorizationCodeAuthorizationURL {
-		const url = new AuthorizationCodeAuthorizationURL(authorizationEndpoint, this.clientId);
-		url.setRedirectURI(this.redirectURI);
-		url.setState(state);
-		url.setS256CodeChallenge(codeVerifier);
+	public createAuthorizationURL(state: string, codeVerifier: string, scopes: string[]): URL {
+		const url = new URL(authorizationEndpoint);
+		url.searchParams.set("client_id", this.clientId);
+		url.searchParams.set("state", state);
+		url.searchParams.set("scope", scopes.join(" "));
+		url.searchParams.set("redirect_uri", this.redirectURI);
+		const codeChallenge = createS256CodeChallenge(codeVerifier);
+		url.searchParams.set("code_challenge_method", "S256");
+		url.searchParams.set("code_challenge", codeChallenge);
 		return url;
 	}
+
 	public async validateAuthorizationCode(
 		code: string,
 		codeVerifier: string
 	): Promise<OAuth2Tokens> {
-		const context = new AuthorizationCodeTokenRequestContext(code);
-		context.setClientId(this.clientId);
-		context.setRedirectURI(this.redirectURI);
-		context.setCodeVerifier(codeVerifier);
-		const tokens = await sendTokenRequest(tokenEndpoint, context);
+		const body = new URLSearchParams();
+		body.set("grant_type", "authorization_code");
+		body.set("code", code);
+		body.set("code_verifier", codeVerifier);
+		body.set("redirect_uri", this.redirectURI);
+		const request = createOAuth2Request(tokenEndpoint, body);
+		const tokens = await sendTokenRequest(request);
 		return tokens;
 	}
 }