+++ src/providers/myanimelist.ts
@@ -1,9 +1,4 @@
-import {
-	AuthorizationCodeAuthorizationURL,
-	AuthorizationCodeTokenRequestContext,
-	RefreshRequestContext
-} from "@oslojs/oauth2";
-import { sendTokenRequest } from "../request.js";
+import { createOAuth2Request, encodeBasicCredentials, sendTokenRequest } from "../request.js";
 
 import type { OAuth2Tokens } from "../oauth2.js";
 
@@ -27,37 +22,43 @@ export class MyAnimeList {
 		this.redirectURI = options?.redirectURI ?? null;
 	}
 
-	public createAuthorizationURL(
-		state: string,
-		codeVerifier: string
-	): AuthorizationCodeAuthorizationURL {
-		const url = new AuthorizationCodeAuthorizationURL(authorizationEndpoint, this.clientId);
+	public createAuthorizationURL(state: string, codeVerifier: string): URL {
+		const url = new URL(authorizationEndpoint);
+		url.searchParams.set("client_id", this.clientId);
+		url.searchParams.set("state", state);
 		if (this.redirectURI !== null) {
-			url.setRedirectURI(this.redirectURI);
+			url.searchParams.set("redirect_uri", this.redirectURI);
 		}
-		url.setState(state);
-		url.setPlainCodeChallenge(codeVerifier);
+		url.searchParams.set("code_challenge_method", "plain");
+		url.searchParams.set("code_challenge", codeVerifier);
 		return url;
 	}
 
-	public async validateAuthorizationCode(
-		code: string,
-		codeVerifier: string
-	): Promise<OAuth2Tokens> {
-		const context = new AuthorizationCodeTokenRequestContext(code);
-		context.authenticateWithHTTPBasicAuth(this.clientId, this.clientSecret);
+	public async validateAuthorizationCode(code: string): Promise<OAuth2Tokens> {
+		const body = new URLSearchParams();
+		body.set("grant_type", "authorization_code");
+		body.set("code", code);
 		if (this.redirectURI !== null) {
-			context.setRedirectURI(this.redirectURI);
+			body.set("redirect_uri", this.redirectURI);
 		}
-		context.setCodeVerifier(codeVerifier);
-		const tokens = await sendTokenRequest(tokenEndpoint, context);
+		const request = createOAuth2Request(tokenEndpoint, body);
+		const encodedCredentials = encodeBasicCredentials(this.clientId, this.clientSecret);
+		request.headers.set("Authorization", `Basic ${encodedCredentials}`);
+		const tokens = await sendTokenRequest(request);
 		return tokens;
 	}
 
 	public async refreshAccessToken(refreshToken: string): Promise<OAuth2Tokens> {
-		const context = new RefreshRequestContext(refreshToken);
-		context.authenticateWithHTTPBasicAuth(this.clientId, this.clientSecret);
-		const tokens = await sendTokenRequest(tokenEndpoint, context);
+		const body = new URLSearchParams();
+		body.set("grant_type", "refresh_token");
+		body.set("refresh_token", refreshToken);
+		if (this.redirectURI !== null) {
+			body.set("redirect_uri", this.redirectURI);
+		}
+		const request = createOAuth2Request(tokenEndpoint, body);
+		const encodedCredentials = encodeBasicCredentials(this.clientId, this.clientSecret);
+		request.headers.set("Authorization", `Basic ${encodedCredentials}`);
+		const tokens = await sendTokenRequest(request);
 		return tokens;
 	}
 }